build: video processing service + Dockerfile
Initial testing phase - first time using multiple
'FROM' statements in a Dockerfile so don't be
surprised if there are issues here.

The inclusion of the test.mp4 file is for testing
the container - I recommend NOT having this in the
project if you are building this - use the CLI to
copy it into a test image instead to reduce clutter.